+++ docker-compose.yml
@@ -53,15 +53,13 @@ services:
   main:
     image: "geos74/uralbur-front:0.0.1"
     ports:
-      - "8080:80"
+      - "3000:80"
     volumes:
-      - ./nginx/main.nginx.conf:/etc/nginx/conf.d/default.conf
+      - ./nginx/front.conf:/etc/nginx/conf.d/default.conf
   cms:
     image: "geos74/uralbur:0.0.1"
-    # ports:
-      # - "3333:80"
     volumes:
-      - ./nginx/cms.nginx.conf:/etc/nginx/conf.d/default.conf
+      - ./nginx/admin.conf:/etc/nginx/conf.d/default.conf
 volumes:
   init-db-mauth:
   db-mauth:
