--- nginx/cms.nginx.conf
@@ -1,16 +1,0 @@
-server {
-  listen 3333;
- 
-  location / {
-    root   /usr/share/nginx/html;
-    try_files $uri /index.html;
-  }
-
-  location /api/mauth/ {
-    proxy_pass http://mauth_app:3001;
-  }
-
-  location /api/mcontent/ {
-    proxy_pass http://mcontent_app:3020;
-  }
-}
