+++ nginx/front.conf
@@ -1,21 +1,9 @@
 server {
   listen 80;
-
+ 
   location / {
     root   /usr/share/nginx/html;
     try_files $uri /index.html;
-  }
-
-  location /admin/ {
-    proxy_pass http://cms:3333;
-  }
-
-  # location ~\.js|css|ico|woff2|txt$ {
-  #    proxy_pass http://cms:3333;
-  # }
-
-  location ~^/static {
-     proxy_pass http://cms:3333;
   }
 
   location /api/mauth/ {
@@ -24,5 +12,17 @@ server {
 
   location /api/mcontent/ {
     proxy_pass http://mcontent_app:3020;
+  }
+
+  location /admin/ {
+    proxy_pass http://uralbur:80;
+  }
+
+  location ~^/static {
+     proxy_pass http://uralbur:80;
+  }
+  
+  location ~^/ionicons {
+     proxy_pass http://uralbur:80;
   }
 }
