+++ src/index.html
@@ -5,6 +5,8 @@
   <title>Ural Bur</title>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+  <meta name="description" content="">
+  <link rel="icon" href="images/logo.svg" type="image/svg+xml" style="width: 32px;">
 
   <link href="https://fonts.googleapis.com/css?family=Montserrat:400,700|Oxygen:400,700" rel="stylesheet">
 
@@ -23,7 +25,8 @@
 </head>
 
 <body>
-  <header role="banner" id="header"></header>
+  <!-- глобальное меню -->
+  <header role="banner" id="navigate"></header>
 
   <div class="top-shadow"></div>
 
@@ -204,6 +207,7 @@
 
   <script src="js/main.js"></script>
   <script src="components/Header/Header.js" type="module"></script>
+  <script src="components/Navigate/Navigate.js" type="module"></script>
   <script src="components/Slider/Slider.js" type="module"></script>
   <script src="components/Solution/Solution.js" type="module"></script>
   <script src="components/Progress/Progress.js" type="module"></script>
